requirements.txt
python-telegram-bot==20.7
groq==0.9.0
chromadb==0.4.22
sentence-transformers==2.2.2
langchain==0.1.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
requests==2.31.0
flask==3.0.0          # для вебхука (опционально)